[hid] -Change: hid_attrib's lng field is not int but long to make sure we have a wider range of numbers available (at least 2^31-1)
git-svn-id: svn://repo.hu/pcb-rnd/trunk@26939 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/hid_attrib.h
@@ -10,7 +10,7 @@
    str_value to the enumeration string.  PCB_HATT_LABEL just shows the
    default str_value. */
 struct pcb_hid_attr_val_s {
-	int lng;
+	long lng;
 	const char *str_value;
 	double real_value;
 	pcb_coord_t coord_value;
